Add manpages and main script
Initial commit to develop branch so that I can branch of develop for
features (if I desire).

The manpages should grow as the script grows, but if I get a better idea
of how I want the UX to be, then I might update the manpages first
(maybe in a separate branch and merging when the feature is
implemented).

ebs-backup currently checks if rsync is installed before continuing,
since maybe future features might break without rsync, and I don't want
to deal with cleanup right now.