Add ssh connection check, volume check, and more
Check if ssh connection is active after an instance runs for some time.
Since I don't want to spam the known_hosts file or deal with the
fingerprint prompt, I use an alias with some options.

An identifiable volume must be present for backup/resotration. I'm just
going to use the "ebs_backup_vol" tag with value "1" to specify that a
volume is both active and ready for use with ebs-backup. The numberical
value might help with weird setups in the future if I decide to support
multiple volumes.

I switched over the instnace to Ubuntu. Rationale:
- Pre-installed lvm support for future-proofing (I might use lvm for
  storage expansion)
- mkfs installed with a variety of supported filesystems, including
  possibly interesting ones like btrfs.
Ubuntu support requires changing the ssh username and ami number.

For now, I've lowered the wait time for ssh, since ubuntu seems fast for
some reason.