Add rough check for space and file limits
Do not backup data if the space exceeds that which is availalbe on the
volume or if the number of files to backup will exceed the limit. The
space limit is based on the current 4096 blocksize on btrfs on the
ubuntu ami, and the file limit is based on btrfs's 2^64 number of files
limit.

Moved termination to a function.

Made rsync ssh quiet.